requirements.txt
PyYAML~=6.0
avro~=1.11.1